refactor(right-panel): create separate folder
